+++ src/main.cpp
@@ -25,6 +25,8 @@
 #endif
 #endif
 
+#include "serial_data_channel/serial_data_manager.h"
+
 #if USE_SDL2
 #include "sdl_renderer/sdl_renderer.h"
 #endif
@@ -66,8 +68,6 @@ int main(int argc, char* argv[]) {
 #if USE_ROS
   std::unique_ptr<rtc::LogSink> log_sink(new ROSLogSink());
   rtc::LogMessage::AddLogToStream(log_sink.get(), rtc::LS_INFO);
-  rtc::scoped_refptr<ROSVideoCapture> capturer(
-      new rtc::RefCountedObject<ROSVideoCapture>(cs));
 #else
   std::unique_ptr<rtc::FileRotatingLogSink> log_sink(
       new rtc::FileRotatingLogSink("./", "webrtc_logs", kDefaultMaxLogFileSize,
@@ -78,21 +78,36 @@ int main(int argc, char* argv[]) {
     return 1;
   }
   rtc::LogMessage::AddLogToStream(log_sink.get(), rtc::LS_INFO);
-  auto size = cs.getSize();
-#if defined(__APPLE__)
-  rtc::scoped_refptr<MacCapturer> capturer = MacCapturer::Create(
-      size.width, size.height, cs.framerate, cs.video_device);
-#elif defined(__linux__)
-  rtc::scoped_refptr<V4L2VideoCapture> capturer = V4L2VideoCapture::Create(cs);
-#else
-  rtc::scoped_refptr<DeviceVideoCapturer> capturer =
-      DeviceVideoCapturer::Create(size.width, size.height, cs.framerate);
 #endif
+
+  auto capturer = ([&]() -> rtc::scoped_refptr<ScalableVideoTrackSource> {
+    if (cs.no_video) {
+      return nullptr;
+    }
+
+#if USE_ROS
+    rtc::scoped_refptr<ROSVideoCapture> capturer(
+        new rtc::RefCountedObject<ROSVideoCapture>(cs));
+#else  // USE_ROS
+    auto size = cs.getSize();
+#if defined(__APPLE__)
+    rtc::scoped_refptr<MacCapturer> capturer = MacCapturer::Create(
+        size.width, size.height, cs.framerate, cs.video_device);
+#elif defined(__linux__)
+    rtc::scoped_refptr<V4L2VideoCapture> capturer =
+        V4L2VideoCapture::Create(cs);
+#else
+    rtc::scoped_refptr<DeviceVideoCapturer> capturer =
+        DeviceVideoCapturer::Create(size.width, size.height, cs.framerate);
+#endif
+#endif  // USE_ROS
+    return capturer;
+  })();
+
   if (!capturer && !cs.no_video) {
     std::cerr << "failed to create capturer" << std::endl;
     return 1;
   }
-#endif
 
 #if USE_SDL2
   std::unique_ptr<SDLRenderer> sdl_renderer = nullptr;
@@ -101,8 +116,8 @@ int main(int argc, char* argv[]) {
         new SDLRenderer(cs.window_width, cs.window_height, cs.fullscreen));
   }
 
-  std::unique_ptr<RTCManager> rtc_manager(
-      new RTCManager(cs, std::move(capturer), sdl_renderer.get()));
+  std::unique_ptr<RTCManager> rtc_manager(new RTCManager(
+      cs, std::move(capturer), sdl_renderer.get()));
 #else
   std::unique_ptr<RTCManager> rtc_manager(
       new RTCManager(cs, std::move(capturer), nullptr));
@@ -110,6 +125,15 @@ int main(int argc, char* argv[]) {
 
   {
     boost::asio::io_context ioc{1};
+
+    std::unique_ptr<RTCDataManager> data_manager = nullptr;
+    if (!cs.serial_device.empty()) {
+      data_manager = SerialDataManager::Create(ioc, cs.serial_device, cs.serial_rate);
+      if (!data_manager) {
+        return 1;
+      }
+      rtc_manager->SetDataManager(data_manager.get());
+    }
 
     boost::asio::signal_set signals(ioc, SIGINT, SIGTERM);
     signals.async_wait(