+++ CHANGES.md
@@ -11,6 +11,34 @@
 
 ## develop
 
+## 2020.2.1
+
+**hotfix**
+
+- [FIX] macOS で --use-sdl オプションを利用すると落ちていたのを修正する
+    - https://bugzilla.libsdl.org/show_bug.cgi?id=4617
+
+## 2020.2
+
+- [UPDATE] CLI11 を v1.9.0 にアップデートする
+    - @voluntas
+- [ADD] Windows 10 対応を追加
+    - @melpon
+- [ADD] Windows の Sora/Ayame モード利用時のシグナリング接続情報に environment / libwebrtc / sora_client を追加
+    - `"environment": "[x64] Windows 10.0 Build 18362"`
+    - `"libwebrtc": "Shiguredo-Build M80.3987@{#2} (80.3987.2.1 fba51dc6)"`
+    - `"sora_client": "WebRTC Native Client Momo 2020.1 (0ff24ff3)"`
+    - @melpon
+- [ADD] ビルド環境を CMake 化
+    - @melpon
+- [CHANGE] ubuntu-18.04_armv8 のビルドを削除
+    - @melpon
+
+## 2020.1
+
+- [UPDATE] libwebrtc を M80.3987@{#2} に上げる
+    - libwebrtc のハッシュは fba51dc69b97f6f170d9c325a38e05ddd69c8b28
+    - @melpon
 - [UPDATE] Momo 2020.1 にバージョンを上げる
     - バージョン番号を <リリース年>.<その年のリリース回数> に変更
     - @voluntas
@@ -24,24 +52,51 @@
     - @melpon
 - [UPDATE] 音声系オプションの --disable-residual-echo-detector を追加する
     - @melpon
+- [ADD] データチャネルを利用したシリアルポートへの読み書き機能を追加する
+    - --serial を指定することでデータチャネル経由でのシリアル読み書きが可能になる
+    - test と ayame モードでのみ利用可能
+    - @tnoho
 - [ADD] 自由に解像度の値を指定できるようにする
     - `--resolution 640x480` のように指定できるようになりました
     - この機能が有効になるのは、カメラに依存するため動作保証はありません
     - @melpon
-- [ADD] Sora 利用時のシグナリング接続情報に enviroment / libwebrtc / sora_client を追加する
+- [ADD] Sora モード利用時のシグナリング接続情報に enviroment / libwebrtc / sora_client を追加する
     - Jetson Nano の場合
         - `"environment": "[aarch64] Ubuntu 18.04.3 LTS (nvidia-l4t-core 32.2.1-20190812212815)"`
-        - `"libwebrtc": "Shiguredo-Build M79@{#5} (79.5.4 b484ec00)"`
+        - `"libwebrtc": "Shiguredo-Build M80.3987@{#2} (80.3987.2.1 15b26e4)"`
         - `"sora_client": "WebRTC Native Client Momo 2020.1 (f6b69e77)"`
     - macOS の場合
         - `"environment": "[x86_64] macOS Version 10.15.2 (Build 19C57)"`
-        - `"libwebrtc": "Shiguredo-Build M79@{#5} (79.5.4 b484ec00)"`
+        - `"libwebrtc": "Shiguredo-Build M80.3987@{#2} (80.3987.2.1 15b26e4)"`
         - `"sora_client": "WebRTC Native Client Momo 2020.1 (f6b69e77)"`
     - Ubuntu 18.04 x86_64 の場合
         - `"environment": "[x86_64] Ubuntu 18.04.3 LTS"`
-        - `"libwebrtc": "Shiguredo-Build M79@{#5} (79.5.4 b484ec00)"`
+        - `"libwebrtc": "Shiguredo-Build M80.3987@{#2} (80.3987.2.1 15b26e4)"`
         - `"sora_client": "WebRTC Native Client Momo 2020.1 (f6b69e77)"`
     - @melpon
+- [ADD] Ayame モード利用時のシグナリング接続情報に enviroment / libwebrtc / ayameClient を追加する
+    - Sora 時の sora_client が ayameClient に変わります
+    - @melpon
+- [ADD] Raspbian ミラーを追加する
+    - @melpon
+- [CHANGE] momo --help の英語化
+    - @shino @msnoigrs
+- [CHANGE] <package>.edit の機能とドキュメントを削除
+    - @melpon
+- [CHANGE] armv6 で SDL を使えなくする
+    - @melpon
+- [FIX] --no-video を指定しているにもかかわらずカメラを一瞬だけ掴むのを修正する
+    - @melpon @mganeko
+- [FIX] SDL が有効でない時に SDL 関連のオプションを指定するとエラーにする
+    - @melpon
+- [FIX] macOS のビルドで Python 2.7 必須を外す
+    - @melpon
+- [FIX] Ayame モードで WebSocket が閉じられた際に再接続処理に進まない箇所を修正
+    - @Hexa
+- [FIX] Ayame モードで シグナリングで bye を受信した際処理として、各 close 処理を追加する
+    - @Hexa
+- [FIX] Ayame モードで 再接続処理の 1 回目を、5 秒後からすぐに実行されるように変更する
+    - @Hexa
 
 ## 19.12.1